refactor: split repository write and profile row helpers
Separate conversation write orchestration, metadata mutation helpers, and session-profile payload/search/hydration builders from their public roots.
